+++ zone/zone_save_state.cpp
@@ -6,6 +6,42 @@
 #include "../common/repositories/zone_state_spawns_repository.h"
 #include "../common/repositories/spawn2_disabled_repository.h"
 
+struct LootEntryStateData {
+	uint32   item_id;
+	uint32_t lootdrop_id;
+
+	// cereal
+	template<class Archive>
+	void serialize(Archive &ar)
+	{
+		ar(
+			CEREAL_NVP(item_id),
+			CEREAL_NVP(lootdrop_id)
+		);
+	}
+};
+
+struct LootStateData {
+	uint32                          copper   = 0;
+	uint32                          silver   = 0;
+	uint32                          gold     = 0;
+	uint32                          platinum = 0;
+	std::vector<LootEntryStateData> entries  = {};
+
+	// cereal
+	template<class Archive>
+	void serialize(Archive &ar)
+	{
+		ar(
+			CEREAL_NVP(copper),
+			CEREAL_NVP(silver),
+			CEREAL_NVP(gold),
+			CEREAL_NVP(platinum),
+			CEREAL_NVP(entries)
+		);
+	}
+};
+
 inline std::string GetLootSerialized(NPC *npc)
 {
 	LootStateData ls         = {};
